Make draw toggle cut off when leaving canvas
Sometimes the 'drag' effect happens on the mouse,
stopping the draw function as you move the mouse.
If you leave the canvas with the mouse button
still held, the mouseup will not register in the
canvas and reentering the canvas without the mouse
button held will cause you to continue to draw.